+++ lib/features/settlement/presentation/screens/confirm_attendance_screen.dart
@@ -1,19 +1,19 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 import 'package:pravo_client/assets/constants.dart';
 import 'package:pravo_client/features/core/presentation/widgets/depth2_app_bar_widget.dart';
 import 'package:pravo_client/features/core/presentation/widgets/primary_button_widget.dart';
-import 'package:pravo_client/features/core/presentation/widgets/vertical_gap_widget.dart';
-import 'package:pravo_client/features/settlement/presentation/widgets/attendance_selector_widget.dart';
+import 'package:pravo_client/features/settlement/presentation/widgets/attendance_list_widget.dart';
 
-class ConfirmAttendanceScreen extends StatelessWidget {
+class ConfirmAttendanceScreen extends ConsumerWidget {
   final int promiseId;
 
   const ConfirmAttendanceScreen({super.key, required this.promiseId});
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     return Scaffold(
       appBar: Depth2AppBarWidget(
         title: '참석 확인',
@@ -23,38 +23,24 @@ class ConfirmAttendanceScreen extends StatelessWidget {
       body: SafeArea(
         child: Column(
           children: [
-            Expanded(
+            const Expanded(
               child: SingleChildScrollView(
                 child: Padding(
                   padding: kScreenPadding,
                   child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      const Text(
+                      Text(
                         '참여자',
                         style: TextStyle(
                           fontSize: 20,
                           fontWeight: FontWeight.w600,
                         ),
                       ),
-                      const SizedBox(
+                      SizedBox(
                         height: 24,
                       ),
-                      Column(
-                        children: [
-                          ListView.separated(
-                            shrinkWrap: true, // <==== limit height. 리스트뷰 크기 고정
-                            primary: false,
-                            itemCount: 10,
-                            itemBuilder: (_, __) =>
-                                const AttendanceSelectorWidget(),
-                            separatorBuilder: (_, __) =>
-                                const VerticalGapWidget(
-                              gapHeight: 12,
-                            ),
-                          ),
-                        ],
-                      ),
+                      AttendanceListWidget(),
                     ],
                   ),
                 ),
